// src/types/enum.ts
/**
 * All Template types for Email
 */
export enum Templates {
  EMAIL_VERIFICATION = "EMAIL_VERIFICATION",
  FORGET_PASSWORD = "FORGET_PASSWORD",
}

export enum DateRange {
  TODAY = "TODAY",
  YEASTERDAY = "YEASTERDAY",
  LAST7DAYS = "LAST7DAYS",
  LAST30DAYS = "LAST30DAYS",
}

export enum Sports {
  FOOTBALL = "FOOTBALL",
  CRICKET = "CRICKET",
}
export enum MatchStatus {
  LIVE = "LIVE",
  FOOTBALL = "FOOTBALL",
}

export enum Games {
  MULTIPLIER = "MULTIPLIER",
  WHEEL = "WHEEL",
  SLOT = "SLOT",
  TOSS = "TOSS",
  DICE = "DICE"
}

export enum UserStatus {
  BANNED = 'BANNED',
  ACTIVE = "ACTIVE"
}

export enum Coin {
  HEADS = "heads",
  TAILS = "tails",
}